fixing stuff after forum report
var/cache is better kept as "keep file content" because omv seems to like doing stuff with it.
Since the only thing that can realistically fill it up is apt's package cache, and you don't seem to like running chronjobs to clean it regularly, I suggest to add a setting file to apt to disable it as explained here http://ask.xmodulo.com/disable-apt-cache-debian-ubuntu.html

Also, the mtab symlinking seems unnecessary, it seems like Debian adopted this symlink as a standard somewhere in 2012, so in wheezy it comes as standard. I did follow older guides, updated the tutorial though.