changing cron-apt options
I doubt that anyone will complain for update speed.
Had a look at what chron-apt is doing.
It's standard config. as here http://www.the-art-of-web.com/system/cron-apt-wheezy/
refreshing lists is ok (it's all in ram), asking it to download packages is nonsense as package cache is disabled. For the sake of safety I'm erasing the (modular) config that asks the package download.
Also using a simpler waiting cycle.

I'm wondering, if this plugin is removed, it would be a good idea to reverse these two settings.
Where can I write the script run when uninstalling?